fix: renamed tsx files to follow PascalCase naming convention for React components.
